Restore delegate_method matcher
Matcher was removed in commit 77dc477c due to issues described in
https://github.com/thoughtbot/shoulda-matchers/issues/252

We've moved away from using Bourne and Mocha's `HaveReceived` matcher and
implemented a `StubbedTarget` class for spying on methods to be delegated.
We also removed the stub call to be handled via `define_method` for
framework-independent support.

Other changes include:

* Moving from `should` syntax to `expect`
* Use 1.9 hash syntax in `delegate_method`